+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v2/LedgerEndService.scala
@@ -1,0 +1,15 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.participant.state.index.v2
+
+import com.digitalasset.ledger.api.domain.LedgerOffset
+
+import scala.concurrent.Future
+
+/**
+  * Serves as a backend to implement ledger end related API calls.
+  **/
+trait LedgerEndService {
+  def currentLedgerEnd(): Future[LedgerOffset.Absolute]
+}