+++ docs/source/support/release-notes.rst
@@ -8,7 +8,16 @@ This page contains release notes for the SDK.
 
 HEAD — ongoing
 --------------
-- Fixed a bug where type check errors would persist if there was a subsequent parse error
+
+DAML Studio
+~~~~~~~~~~~
+
+- Fixed a bug where type check errors would persist if there was a subsequent parse error.
+
+DAML Compiler
+~~~~~~~~~~~~~
+
+- **BREAKING CHANGE**: Drop support for DAML-LF 1.2. Compiling to DAML-LF 1.3 should work without any code changes, although we highly recommend not specifying a target DAML-LF version at all.
 
 SQL Extractor
 ~~~~~~~~~~~~~